+++ Code Mods/VSTutorial - 6 - Simple GUI & Hotkeys Tutorial/VSTutorial/VSTutorialModSystem.cs
@@ -1,4 +1,4 @@
-﻿using AnnoyingTextSystem;
+﻿using VSTutorial.GUI;
 using Vintagestory.API.Client;
 using Vintagestory.API.Common;
