+++ Code Mods/VSTutorial - 7 - Basic Harmony Patching/VSTutorial.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34330.188
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSTutorial", "VSTutorial\VSTutorial.csproj", "{EF5AB32D-83B1-4C08-A5B6-DBCD509C3B59}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CakeBuild", "ZZCakeBuild\CakeBuild.csproj", "{5828ECF7-9ED7-4FA5-874E-4DC8787CEDFF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EF5AB32D-83B1-4C08-A5B6-DBCD509C3B59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF5AB32D-83B1-4C08-A5B6-DBCD509C3B59}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF5AB32D-83B1-4C08-A5B6-DBCD509C3B59}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF5AB32D-83B1-4C08-A5B6-DBCD509C3B59}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5828ECF7-9ED7-4FA5-874E-4DC8787CEDFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5828ECF7-9ED7-4FA5-874E-4DC8787CEDFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5828ECF7-9ED7-4FA5-874E-4DC8787CEDFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5828ECF7-9ED7-4FA5-874E-4DC8787CEDFF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {848437EF-598F-48F2-BAC1-04562796FA2B}
+	EndGlobalSection
+EndGlobal
