+++ Code Mods/VSTutorial - 7 - Basic Harmony Patching/VSTutorial/VSTutorial.csproj
