+++ Code Mods/VSTutorial - 7 - Basic Harmony Patching/ZZCakeBuild/CakeBuild.csproj
