+++ Code Mods/VSTutorial - 7 - Basic Harmony Patching/build.ps1
@@ -1,0 +1,2 @@
+dotnet run --project ZZCakeBuild/CakeBuild.csproj -- $args
+exit $LASTEXITCODE;
