+++ Code Mods/VSTutorial - 7 - Basic Harmony Patching/build.sh
@@ -1,0 +1,1 @@
+dotnet run --project ./ZZCakeBuild/CakeBuild.csproj -- "$@"
