--- code_mods/AnnoyingTextSystem/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "dotnet.defaultSolution": "AnnoyingTextSystem.sln"
-}
