--- code_mods/AnnoyingTextSystem/.vscode/tasks.json
@@ -1,28 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "build",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "build",
-                "-c",
-                "Debug",
-                "${workspaceFolder}/AnnoyingTextSystem/AnnoyingTextSystem.csproj"
-            ],
-            "problemMatcher": "$msCompile"
-        },
-        {
-            "label": "package",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "run",
-                "--project",
-                "${workspaceFolder}/CakeBuild/CakeBuild.csproj"
-            ],
-            "problemMatcher": "$msCompile"
-        }
-    ]
-}
