--- code_mods/AnnoyingTextSystem/AnnoyingTextSystem/GuiDialogAnnoyingText.cs
@@ -1,40 +1,0 @@
-using Vintagestory.API.Client;
-
-namespace AnnoyingTextSystem
-{
-    public class GuiDialogAnnoyingText : GuiDialog
-    {
-        public override string ToggleKeyCombinationCode => "annoyingtextgui";
-
-        public GuiDialogAnnoyingText(ICoreClientAPI capi) : base(capi)
-        {
-            SetupDialog();
-        }
-
-        private void SetupDialog()
-        {
-            // Auto-sized dialog at the center of the screen
-            ElementBounds dialogBounds = ElementStdBounds.AutosizedMainDialog.WithAlignment(EnumDialogArea.CenterMiddle);
-
-            // Just a simple 300x300 pixel box
-            ElementBounds textBounds = ElementBounds.Fixed(0, 40, 300, 100);
-
-            // Background boundaries. Again, just make it fit it's child elements, then add the text as a child element
-            ElementBounds bgBounds = ElementBounds.Fill.WithFixedPadding(GuiStyle.ElementToDialogPadding);
-            bgBounds.BothSizing = ElementSizing.FitToChildren;
-            bgBounds.WithChildren(textBounds);
-
-            SingleComposer = capi.Gui.CreateCompo("myAwesomeDialog", dialogBounds)
-                .AddShadedDialogBG(bgBounds)
-                .AddDialogTitleBar("Heck yeah!", null)
-                .AddStaticText("This is a piece of text at the center of your screen - Enjoy!", CairoFont.WhiteDetailText(), textBounds)
-              .Compose()
-            ;
-        }
-
-        private void OnTitleBarCloseClicked()
-        {
-            TryClose();
-        }
-    }
-}
