--- code_mods/AnnoyingTextSystem/AnnoyingTextSystem/Properties/launchSettings.json
@@ -1,16 +1,0 @@
-{
-  "profiles": {
-    "Client": {
-      "commandName": "Executable",
-      "executablePath": "$(VINTAGE_STORY)/Vintagestory.exe",
-      "commandLineArgs": "--tracelog --addModPath \"$(ProjectDir)/bin/$(Configuration)/Mods\"",
-      "workingDirectory": "$(VINTAGE_STORY)"
-    },
-    "Server": {
-      "commandName": "Executable",
-      "executablePath": "$(VINTAGE_STORY)/VintagestoryServer.exe",
-      "commandLineArgs": "--tracelog --addModPath \"$(ProjectDir)/bin/$(Configuration)/Mods\"",
-      "workingDirectory": "$(VINTAGE_STORY)"
-    }
-  }
-}
