--- code_mods/AnnoyingTextSystem/AnnoyingTextSystem/modinfo.json
@@ -1,13 +1,0 @@
-{
-    "type": "code",
-    "modid": "annoyingtextsystem",
-    "name": "AnnoyingTextSystem",
-    "authors": ["Tyron"],
-    "description": "",
-    "version": "1.0.0",
-    "side" : "Client",
-    "dependency": {
-      "game": ""
-    }
-  }
-
