--- code_mods/AnnoyingTextSystem/build.ps1
@@ -1,2 +1,0 @@
-dotnet run --project CakeBuild/CakeBuild.csproj -- $args
-exit $LASTEXITCODE;
