--- code_mods/AnnoyingTextSystem/build.sh
@@ -1,1 +1,0 @@
-dotnet run --project ./CakeBuild/CakeBuild.csproj -- "$@"
